refactor : ImgInput 재사용하지 않는 컴포넌트 분리
