+++ front/src/components/index.ts
@@ -1,6 +1,5 @@
 export { default as TextInput } from './TextInput';
 export { default as SearchBar } from './SearchBar';
-export { default as ImgInput } from './ImgInput';
 export { default as MakeWorldcupForm } from './MakeWorldcupForm';
 export { default as ImgPreView } from './ImgPreView';
 export { default as ImgPreViewList } from './ImgPreViewList';
